[seller] refactor(223): 파일명 오타 수정
