+++ apps/seller/src/features/products/create/create-preview/create-preview-slider.hook.ts
