+++ apps/seller/src/features/products/create/create-preview/index.ts
@@ -1,4 +1,4 @@
-export { useCreatePreviewHook } from './create-preivew.hook'
+export { useCreatePreviewHook } from './create-preview.hook'
 export { CreatePreviewOptionItemUi } from './create-preview-option-item.ui'
 export { ProductPreviewModal } from './create-preview-modal.ui'
-export { usePreviewSlider } from './create-preivew-slider.hook'
+export { usePreviewSlider } from './create-preview-slider.hook'
